Added nightly upsample tests
